# docs/openapi.yaml
openapi: 3.0.3
info:
  title: BeautyBook API
  version: 1.0.0
  description: |
    Especificacao OpenAPI da API REST do BeautyBook, baseada nos epicos, user stories
    e regras de negocio do projeto.

    Principais regras refletidas nesta API:
    - Autenticacao obrigatoria para acesso aos recursos protegidos
    - Controle de acesso por perfil (`ADMIN` e `CLIENT`)
    - Gestao de servicos com regras condicionais para `necessitaAvaliacao`
    - Configuracao de expediente, dias de funcionamento e intervalos
    - Consulta de disponibilidade considerando agenda, sobreposicao e duracao do servico
    - Criacao e cancelamento de agendamentos com regras operacionais do salao
servers:
  - url: http://localhost:3000/api
    description: Ambiente local
...
security:
  - bearerAuth: []
paths:
...
  /business-rules:
    get:
      tags: [Configuracao Operacional]
      summary: Consulta configuracao operacional do salao
      description: Retorna horario de funcionamento, dias de operacao e intervalos.
      responses:
        '200':
          description: Configuracao retornada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRules'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags: [Configuracao Operacional]
      summary: Atualiza configuracao operacional do salao
      description: Recurso restrito ao perfil `ADMIN`.
      x-roles:
        - ADMIN
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BusinessRulesUpsertRequest'
      responses:
        '200':
          description: Configuracao atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRules'
        '400':
          description: Dados invalidos para configuracao operacional
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
              examples:
                invalidOpeningTime:
                  summary: horaAbertura fora do formato
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: horaAbertura
                        message: Formato de hora invalido. Use HH:MM
                invalidClosingTime:
                  summary: horaFechamento fora do formato
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: horaFechamento
                        message: Formato de hora invalido. Use HH:MM
                invalidTimeOrder:
                  summary: Hora final menor ou igual a inicial
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: horaFechamento
                        message: horaFechamento deve ser maior que horaAbertura
                invalidWorkingDay:
                  summary: Dia da semana invalido
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: diasFuncionamento
                        message: diasFuncionamento aceita apenas dias validos entre SUNDAY e SATURDAY
                invalidBreakFormat:
                  summary: Intervalo com formato invalido
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: intervalos[0].horaInicio
                        message: Formato de hora invalido. Use HH:MM
                      - field: intervalos[0].horaFim
                        message: Formato de hora invalido. Use HH:MM
                invalidBreakOrder:
                  summary: Hora final do intervalo menor ou igual a inicial
                  value:
                    code: VALIDATION_ERROR
                    message: Configuracao operacional invalida
                    details:
                      - field: intervalos[0]
                        message: horaFim deve ser maior que horaInicio
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /appointments:
    post:
      tags: [Agendamentos]
      summary: Cria um agendamento ou solicitacao de avaliacao
      description: |
        Recurso restrito ao perfil `CLIENT`. Quando o servico possuir
        `necessitaAvaliacao = true`, a API cria uma solicitacao com status
        `PENDENTE_AVALIACAO` e nao bloqueia horario automaticamente.
      x-roles:
        - CLIENT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '201':
          description: Agendamento ou solicitacao criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentCreateResponse'
        '400':
          description: Requisicao invalida ou regra de negocio violada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalidPhone:
                  summary: Telefone invalido
                  value:
                    code: INVALID_PHONE
                    message: O telefone deve possuir 11 digitos
                invalidSchedule:
                  summary: Fora do expediente
                  value:
                    code: OUTSIDE_BUSINESS_HOURS
                    message: O horario informado esta fora do expediente
                pastDateTime:
                  summary: Data ou horario no passado
                  value:
                    code: APPOINTMENT_IN_PAST
                    message: Nao e permitido criar agendamentos em datas ou horarios passados
                appointmentLimit:
                  summary: Limite diario excedido
                  value:
                    code: DAILY_APPOINTMENT_LIMIT_REACHED
                    message: O cliente ja possui 2 agendamentos nesta data
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '409':
          description: Conflito de horario ou sobreposicao
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                exactConflict:
                  value:
                    code: TIME_SLOT_UNAVAILABLE
                    message: Ja existe um agendamento para este horario
                overlapConflict:
                  value:
                    code: TIME_SLOT_OVERLAP
                    message: O horario solicitado conflita com outro agendamento existente

    get:
      tags: [Agendamentos]
      summary: Lista a agenda completa por data
      description: Recurso restrito ao perfil `ADMIN`.
      x-roles:
        - ADMIN
      parameters:
        - name: data
          in: query
          required: true
          description: Data no formato ISO `YYYY-MM-DD`.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Agenda do dia retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminAppointmentView'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /appointments/{appointmentId}:
    get:
      tags: [Agendamentos]
      summary: Consulta um agendamento por identificador
      description: |
        `CLIENT` pode consultar apenas seus proprios agendamentos.
        `ADMIN` pode consultar qualquer agendamento.
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      responses:
        '200':
          description: Agendamento encontrado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /appointments/{appointmentId}/cancel:
    patch:
      tags: [Agendamentos]
      summary: Cancela um agendamento
      description: |
        `CLIENT` pode cancelar apenas seus proprios agendamentos.
        `ADMIN` pode cancelar qualquer agendamento.
        O cancelamento exige antecedencia minima de 48 horas.
      parameters:
        - $ref: '#/components/parameters/AppointmentId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCancellationRequest'
      responses:
        '200':
          description: Agendamento cancelado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Appointment'
        '400':
          description: Cancelamento nao permitido pela regra de negocio
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: CANCELLATION_WINDOW_EXPIRED
                message: Cancelamento permitido apenas com 48 horas de antecedencia
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    AppointmentId:
      name: appointmentId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    UnauthorizedError:
      description: Usuario nao autenticado ou token invalido
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: UNAUTHORIZED
            message: Autenticacao obrigatoria
    ForbiddenError:
      description: Usuario autenticado sem permissao para o recurso
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: FORBIDDEN
            message: Voce nao possui permissao para acessar este recurso
    NotFoundError:
      description: Recurso nao encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: RESOURCE_NOT_FOUND
            message: Recurso nao encontrado
    ValidationError:
      description: Dados invalidos na requisicao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationErrorResponse'
  schemas:
...
    BusinessRules:
      type: object
      required:
        - horaAbertura
        - horaFechamento
        - diasFuncionamento
        - intervalos
      properties:
        horaAbertura:
          $ref: '#/components/schemas/LocalTime'
        horaFechamento:
          $ref: '#/components/schemas/LocalTime'
        diasFuncionamento:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/WorkingDay'
          example:
            - TUESDAY
            - WEDNESDAY
            - THURSDAY
            - FRIDAY
            - SATURDAY
        intervalos:
          type: array
          items:
            $ref: '#/components/schemas/BreakInterval'
    BusinessRulesUpsertRequest:
      allOf:
        - $ref: '#/components/schemas/BusinessRules'
    WorkingDay:
      type: string
      enum:
        - SUNDAY
        - MONDAY
        - TUESDAY
        - WEDNESDAY
        - THURSDAY
        - FRIDAY
        - SATURDAY
    BreakInterval:
      type: object
      required:
        - horaInicio
        - horaFim
      properties:
        horaInicio:
          $ref: '#/components/schemas/LocalTime'
        horaFim:
          $ref: '#/components/schemas/LocalTime'
...
    AppointmentCreateRequest:
      type: object
      required:
        - nomeCliente
        - nomeServico
        - data
        - hora
        - telefone
      properties:
        nomeCliente:
          type: string
          example: Cliente Exemplo
        nomeServico:
          type: string
          example: Corte feminino
        data:
          type: string
          format: date
          description: Data no padrao ISO `YYYY-MM-DD`.
        hora:
          $ref: '#/components/schemas/LocalTime'
        telefone:
          type: string
          pattern: '^\d{11}$'
          description: Telefone com 11 digitos (DDD + numero).
          example: '11999998888'
        observacao:
          type: string
          nullable: true
          maxLength: 500
    AppointmentCreateResponse:
      type: object
      required:
        - id
        - status
        - nomeCliente
        - nomeServico
        - data
        - hora
        - telefone
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        nomeCliente:
          type: string
          example: Cliente Exemplo
        nomeServico:
          type: string
          example: Corte feminino
        data:
          type: string
          format: date
        hora:
          $ref: '#/components/schemas/LocalTime'
        telefone:
          type: string
        mensagem:
          type: string
          nullable: true
          example: Este servico requer avaliacao. Entre em contato com o salao.
        tempoEstimado:
          allOf:
            - $ref: '#/components/schemas/LocalTime'
          nullable: true
        valor:
          type: number
          format: double
          nullable: true
        createdAt:
          type: string
          format: date-time
    Appointment:
      type: object
      required:
        - id
        - nomeCliente
        - nomeServico
        - status
        - data
        - hora
        - telefone
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        nomeCliente:
          type: string
          example: Cliente Exemplo
        nomeServico:
          type: string
          example: Corte feminino
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        data:
          type: string
          format: date
        hora:
          $ref: '#/components/schemas/LocalTime'
        telefone:
          type: string
        observacao:
          type: string
          nullable: true
        motivoCancelamento:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        cancelledAt:
          type: string
          format: date-time
          nullable: true
    AdminAppointmentView:
      type: object
      required:
        - id
        - nomeCliente
        - nomeServico
        - status
        - data
        - hora
      properties:
        id:
          type: string
          format: uuid
        nomeCliente:
          type: string
          example: Ana Paula
        nomeServico:
          type: string
          example: Escova
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        data:
          type: string
          format: date
        hora:
          $ref: '#/components/schemas/LocalTime'
        valor:
          type: number
          format: double
          nullable: true
          example: 80.0
    AppointmentCancellationRequest:
      type: object
      properties:
        reason:
          type: string
          maxLength: 255
          nullable: true
    AppointmentStatus:
      type: string
      enum:
        - CONFIRMED
        - PENDENTE_AVALIACAO
        - CANCELLED
    LocalTime:
      type: string
      pattern: '^([01]\d|2[0-3]):[0-5]\d$'
      example: '13:30'
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: BUSINESS_RULE_VIOLATION
        message:
          type: string
          example: Nao foi possivel concluir a operacao
    ValidationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            details:
              type: array
              items:
                $ref: '#/components/schemas/FieldValidationError'
    FieldValidationError:
      type: object
      required:
        - field
        - message
      properties:
        field:
          type: string
          example: tempoServico
        message:
          type: string
          example: Formato de hora invalido. Use HH:MM